% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_robex.R
\name{install_robex}
\alias{install_robex}
\title{Install ROBEX tools}
\usage{
install_robex(...)
}
\arguments{
\item{...}{arguments to pass to \code{\link{system.file}}}
}
\value{
Logical of success/failure
}
\description{
Install ROBEX tools
}
