% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robex.R
\name{robex}
\alias{robex}
\title{Executes Robust Brain Extraction}
\usage{
robex(infile, outfile = tempfile(fileext = ".nii.gz"), verbose = TRUE,
  ...)
}
\arguments{
\item{infile}{Input filename}

\item{outfile}{Output skull-stripped file}

\item{verbose}{print diagnostic messages}

\item{...}{arguments to be passed to \code{\link{system}}}
}
\value{
List of result of \code{system} command,
names of input and output file
}
\description{
Executes Robust Brain Extraction
}
\examples{
result = robex(infile = "", outfile = "")

if (requireNamespace("kirby21.t1", quietly = TRUE)) {
   infile = kirby21.t1::get_t1_filenames(id = "113", visit = 1)
   if (is.null(infile)) {
      infile = ""
   }
   if (!file.exists(infile)) {
      outdir = tempdir()
      try({kirby21.t1::download_t1_data(outdir = outdir)})
     infile = kirby21.t1::get_t1_filenames(
     id = "113", visit = 1,
     outdir = outdir)
   }
   if (!is.null(infile)) {
     if (file.exists(infile)) {
       result = robex(infile = infile)
       stopifnot(file.exists(result$outfile))
     }
   }
}
}
