% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robex_cmd.R
\name{robex_cmd}
\alias{robex_cmd}
\title{Find ROBEX Executable}
\usage{
robex_cmd(...)
}
\arguments{
\item{...}{arguments to pass to \code{\link{system.file}}}
}
\value{
Character vector of path
}
\description{
Returns a path for the ROBEX executable
}
